requirements.txt
Flask==2.2.5
tensorflow==2.10.0
Pillow==9.4.0
opencv-python==4.8.0.74
google-generativeai==0.1.0
tk==0.1.0
numpy==1.23.5
